requirements.txt
datasets==4.8.4
huggingface_hub==1.7.2
matplotlib==3.10.8
numpy==1.26.4
pandas==3.0.1
pillow==12.0.0
torch==2.11.0
torch-fidelity==0.3.0
torchmetrics==1.8.2
torchvision==0.26.0
tqdm
